+++ htdocs/js/features.js
@@ -14,26 +14,45 @@ function setFeatures() {
     features.featCacert = settings.features.featCacert;
     features.featHome = settings.webuiSettings.enableHome;
     features.featVolumeLevel = settings.webuiSettings.footerVolumeLevel;
-    features.featLocalPlayback = settings.webuiSettings.enableLocalPlayback === true ?
-        (settings.partition.mpdStreamPort > 0 || settings.partition.streamUri.length > 0 ? true : false) : false;
-    features.featScripting = settings.webuiSettings.enableScripting === true ?
-        settings.features.featScripting : false;
+    features.featLocalPlayback = settings.webuiSettings.enableLocalPlayback === true
+        ? settings.partition.mpdStreamPort > 0 || settings.partition.streamUri.length > 0
+            ? true
+            : false
+        : false;
+    features.featScripting = settings.webuiSettings.enableScripting === true
+        ? settings.features.featScripting
+        : false;
     features.featTimer = settings.webuiSettings.enableTimer;
     features.featTrigger = settings.webuiSettings.enableTrigger;
     features.featMediaSession = checkMediaSessionSupport();
     features.featFooterNotifications = settings.webuiSettings.footerNotifications;
     features.featSession = settings.pin;
+    features.featAlbums = settings.features.featTags === true
+        ? settings.features.featAlbums
+        : false;
 
     //mpd features
     if (settings.partition.mpdConnected === true) {
         features.featLibrary = settings.features.featLibrary;
-        features.featLyrics = settings.webuiSettings.enableLyrics === true ? settings.features.featLibrary : false;
-        features.featMounts = settings.webuiSettings.enableMounts === true ? settings.features.featMounts : false;
-        features.featNeighbors = settings.webuiSettings.enableMounts === true ? settings.features.featNeighbors : false;
-        features.featPartitions = settings.webuiSettings.enablePartitions === true ? settings.features.featPartitions : false;
+        features.featLyrics = settings.webuiSettings.enableLyrics === true
+            ? settings.features.featLibrary
+            : false;
+        features.featMounts = settings.webuiSettings.enableMounts === true
+            ? settings.features.featMounts
+            : false;
+        features.featNeighbors = settings.webuiSettings.enableMounts === true
+            ? settings.features.featNeighbors
+            : false;
+        features.featPartitions = settings.webuiSettings.enablePartitions === true
+            ? settings.features.featPartitions
+            : false;
         features.featPlaylists = settings.features.featPlaylists;
-        features.featSmartplsAvailable = settings.features.featPlaylists === true && settings.features.featTags === true ? settings.smartpls : false;
-        features.featSmartpls = settings.features.featPlaylists === true && settings.features.featTags === true ? settings.smartpls : false;
+        features.featSmartplsAvailable = settings.features.featPlaylists === true && settings.features.featTags === true
+            ? settings.smartpls
+            : false;
+        features.featSmartpls = settings.features.featPlaylists === true && settings.features.featTags === true
+            ? settings.smartpls
+            : false;
         features.featStickers = settings.features.featStickers;
         features.featTags = settings.features.featTags;
         features.featBinarylimit = settings.features.featBinarylimit;