+++ src/mpd_client/features.c
@@ -124,7 +124,7 @@ static void features_commands(struct t_partition_state *partition_state) {
         while ((pair = mpd_recv_command_pair(partition_state->conn)) != NULL) {
             if (strcmp(pair->value, "sticker") == 0) {
                 MYMPD_LOG_DEBUG(partition_state->name, "MPD supports stickers");
-                partition_state->mpd_state->feat_stickers = true;
+                partition_state->mpd_state->feat_stickers = partition_state->mympd_state->config->stickers;
             }
             else if (strcmp(pair->value, "listplaylists") == 0) {
                 MYMPD_LOG_DEBUG(partition_state->name, "MPD supports playlists");