+++ src/mpd_worker/cache.c
@@ -183,7 +183,8 @@ static bool cache_init(struct t_mpd_worker_state *mpd_worker_state, rax *album_c
     #endif
     const bool create_album_cache = mpd_worker_state->partition_state->mpd_state->feat_tags &&
         mpd_client_tag_exists(&mpd_worker_state->partition_state->mpd_state->tags_mympd, MPD_TAG_ALBUM) &&
-        mpd_client_tag_exists(&mpd_worker_state->partition_state->mpd_state->tags_mympd, mpd_worker_state->partition_state->mpd_state->tag_albumartist);
+        mpd_client_tag_exists(&mpd_worker_state->partition_state->mpd_state->tags_mympd, mpd_worker_state->partition_state->mpd_state->tag_albumartist) &&
+        mpd_worker_state->config->albums;
     sds key = sdsempty();
     do {
         if (mpd_search_db_songs(mpd_worker_state->partition_state->conn, false) == false ||
@@ -200,7 +201,7 @@ static bool cache_init(struct t_mpd_worker_state *mpd_worker_state, rax *album_c
                 MYMPD_LOG_NOTICE("default", "Skipping album cache creation, (Album)Artist and Album tags must be enabled");
             }
             while ((song = mpd_recv_song(mpd_worker_state->partition_state->conn)) != NULL) {
-                //sticker cache
+                // sticker cache
                 if (mpd_worker_state->partition_state->mpd_state->feat_stickers == true) {
                     const char *uri = mpd_song_get_uri(song);
                     struct t_sticker *sticker = malloc_assert(sizeof(struct t_sticker));