+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Describe the release here.
 ## Changelog
 
 - Feat: Use mpack to store caches on disc #1104
+- Feat: Add a switch to disable album and sticker features #1103
 - Upd: remove default smart playlist generation tag
 
 ***